% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_4s_calc.R
\name{distance_4s_calc}
\alias{distance_4s_calc}
\title{Calculate distance covered within 4s after odor encounter.}
\usage{
distance_4s_calc(object)
}
\arguments{
\item{object}{a \code{data.frame} as produced by \code{read.esm}}
}
\value{
a \code{data.frame} containing net replacement within 4s per fly and odor
}
\description{
\code{distance_4s_calc} calculates the distance covered within 4s after odor encounter.
Rows containing missing values in the first 4s after odor encounter will be removed.
Results are written to \code{responses.csv}.
}

