% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanwithin.all.1to7.R
\name{mean.within.all.1to7}
\alias{mean.within.all.1to7}
\title{Calculate mean response time-course per fly and odor.}
\usage{
\method{mean}{within.all.1to7}(object)
}
\arguments{
\item{object}{a \code{data.frame} as produced by \code{read.esm}}
}
\value{
a \code{data.frame} containing mean response time-courses per fly and odor
}
\description{
\code{mean.within.all.1to7} calculates the mean response time-course per fly and odor.
Rows containing missing values in the interval between 1s before and 7s after odor encounter will be removed.
Results are written to \code{within_fly_mean_traces.csv} and a heatmap respresentation is exported (\code{mean_speed_heatmap.pdf}).
}

