% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_activity_all_median_1to7_R.R
\name{plot_activity_all_median_1to7}
\alias{plot_activity_all_median_1to7}
\title{Calculate and plot median activity across flies.}
\usage{
plot_activity_all_median_1to7(object)
}
\arguments{
\item{object}{a \code{data.frame} as produced by \code{read.esm}}
}
\value{
a \code{data.frame} containing mean activity time-courses per fly and odor
}
\description{
\code{plot_activity_all_median_1to7} calculates the mean activity within flies and calculates and plots the median activity time-courses across flies around odor encounter.
Rows containing missing values in the interval between 1s before and 7s after odor encounter will be removed.
Activity timecourses for individual flies are written to \code{within_fly_activity.csv} and a overview plot is exported (\code{activity_overview.pdf}).
}

