% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box_all_1to7.R
\name{plot_box_all_1to7}
\alias{plot_box_all_1to7}
\title{Plot overview over the experiment across flies.}
\usage{
plot_box_all_1to7(object)
}
\arguments{
\item{object}{object a \code{data.frame} as produced by \code{read.esm}}
}
\value{
\code{speed_overview.pdf}
}
\description{
Plot overview over the experiment across flies. First calculates median response time-courses per odorant within flies.
From these, a boxplot including lines for mean and median time-courses per odorant across flies is created.
}

