% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace_all_1to7.R
\name{plot_trace_all_1to7}
\alias{plot_trace_all_1to7}
\title{Plot overview over the experiment within flies.}
\usage{
plot_trace_all_1to7(object)
}
\arguments{
\item{object}{object a \code{data.frame} as produced by \code{read.esm}.}
}
\value{
one \code{_overview.pdf} per fly analyzed.
}
\description{
Plot overview over the experiment within flies. Creates a \code{_overview.pdf} per fly in a given experiment containing one line plot including
all single tracking events and the fly's mean and median response time-course.
}

