% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/read.esm.R
=======
% Please edit documentation in R/read_esm.R
>>>>>>> 3f25a7675a734d93be25fde6369d8f27e3da2465
\name{read_esm}
\alias{read_esm}
\title{Read all esm-csv in a gven directory.}
\usage{
<<<<<<< HEAD
read_esm(directory, header = TRUE, delimiter = ";", parser = "_")
}
\description{
Will read all esm files in the given directory and concatenate them in single \code{data.frame}. Filenames will be used to keep track of individual flies in different experiments carrying the same within-experiment number.
}

=======
read_esm(directory, delimiter = ";", parser = "_")
}
\arguments{
\item{directory}{directory containing extended speed matrizes}

\item{delimiter}{delimiter used in extended speed matrizes, defaults to ";"
(output from German MS Excel)}

\item{parser}{pattern to parse file names for naming of individual flies}
}
\value{
\code{data.frame} containing combined data
}
\description{
Will read all esm files in the given directory and concatenate them in single
\code{data.frame}. Filenames will be used to keep track of individual flies
in different experiments carrying the same within-experiment number.
}
>>>>>>> 3f25a7675a734d93be25fde6369d8f27e3da2465
