% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_per_pulse.R
\name{response_per_pulse}
\alias{response_per_pulse}
\title{Calculate mean net displacement per odor pulse.}
\usage{
response_per_pulse(object)
}
\arguments{
\item{object}{a \code{data.frame} as produced by \code{read.esm}}
}
\value{
a \code{data.frame}
}
\description{
Calculate mean response per odor pulse in a given experiment. Can be used to check data for adaptation/habituation effects.
Response is defined as the net displacement within 4s after odor encounter.
}

