% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.esm.R
\name{read_esm}
\alias{read_esm}
\title{Read all esm-csv in a gven directory.}
\usage{
read_esm(directory, header = TRUE, delimiter = ";", parser = "_")
}
\description{
Will read all esm files in the given directory and concatenate them in single \code{data.frame}. Filenames will be used to keep track of individual flies in different experiments carrying the same within-experiment number.
}

